test: 요구사항 테스트 추가
